requirements.txt
Django==1.7
Markdown==2.4
Pillow==2.6.1
dj-database-url==0.3.0
dj-static==0.0.6
django-admin-bootstrapped==0.9
django-cors-headers==0.12
django-filter==0.8
djangorestframework==2.3.13
gdata==2.0.18
gunicorn==18.0
openpyxl==1.8.5
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
